+++ R/ggcorrmat.R
@@ -108,7 +108,7 @@
 #' \url{https://indrajeetpatil.github.io/ggstatsplot/articles/web_only/ggcorrmat.html}
 #'
 #' @examples
-#'
+#' \donttest{
 #' # for reproducibility
 #' set.seed(123)
 #'
@@ -161,6 +161,7 @@
 #'   outline.col = "white",
 #'   title = "Dataset: Iris"
 #' )
+#' }
 #' @export
 
 # defining the function
@@ -266,13 +267,8 @@ ggcorrmat <- function(data,
   # ===================== statistics ========================================
   #
   if (corr.method %in% c("pearson", "spearman", "kendall")) {
-    # confidence interval computation can take some time and also produce
-    # warnings, so compute them only when requested by the user
-    if (output == "ci") {
-      ci <- TRUE
-    } else {
-      ci <- FALSE
-    }
+    # compute confidence intervals only when requested by the user
+    ci <- ifelse(test = output == "ci", yes = TRUE, no = FALSE)
 
     # computing correlations using `psych` package
     corr_df <-
@@ -460,8 +456,10 @@ ggcorrmat <- function(data,
           ylab = NULL
         )
 
-      # adding ggstatsplot theme for correlation matrix
-      if (isTRUE(ggstatsplot.layer)) plot <- plot + theme_corrmat()
+      # adding `ggstatsplot` theme for correlation matrix
+      if (isTRUE(ggstatsplot.layer)) {
+        plot <- plot + theme_corrmat()
+      }
     }
   }