+++ R/global_vars.R
@@ -204,7 +204,12 @@ utils::globalVariables(
     "pair1",
     "W",
     "Parameter",
-    "p.value.formatted"
+    "p.value.formatted",
+    "n_per_level",
+    "mean...summary",
+    "mean.conf.low...summary",
+    "mean.conf.high...summary",
+    "n...summary"
   ),
   package = "ggstatsplot",
   add = FALSE