+++ R/grouped_ggbarstats.R
@@ -23,7 +23,6 @@
 #' @inherit ggbarstats return return
 #'
 #' @examples
-#'
 #' \donttest{
 #' # with condition and with count data
 #' library(jmv)