+++ R/grouped_ggbetweenstats.R
@@ -56,7 +56,6 @@
 #'   messages = FALSE
 #' )
 #' }
-#'
 #' @export
 
 # defining the function