+++ R/helpers_gghistostats_graphics.R
@@ -135,7 +135,6 @@ line_labeler <- line_labeller
 #'   test.value.line = TRUE
 #' )
 #' }
-#'
 #' @keywords internal
 
 # function body