+++ inst/doc/additional.R
@@ -1,8 +1,8 @@
-## ----library, include = FALSE--------------------------------------------
+## ----library, include = FALSE-------------------------------------------------
 library(rlang)
 library(ggstatsplot)
 
-## ----setup, include = FALSE----------------------------------------------
+## ----setup, include = FALSE---------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"